Task 30 : Redefine duration as a double instead of long in RouteDTO and Route entity

+++ src/main/java/com/springshortpath/app/model/Route.java
@@ -35,5 +35,5 @@ public class Route {
     private String arriveTime;
 
     @Property
-    private Long duration;
+    private Double duration;
 }
